Move unit tests to the correct files
