+++ cnd/src/config/mod.rs
@@ -41,3 +41,82 @@ pub struct Ethereum {
     #[serde(with = "url_serde")]
     pub node_url: reqwest::Url,
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use reqwest::Url;
+
+    #[test]
+    fn bitcoin_deserializes_correctly() {
+        let file_contents = vec![
+            r#"
+            network = "mainnet"
+            node_url = "http://example.com"
+            "#,
+            r#"
+            network = "testnet"
+            node_url = "http://example.com"
+            "#,
+            r#"
+            network = "regtest"
+            node_url = "http://example.com"
+            "#,
+        ];
+
+        let expected = vec![
+            Bitcoin {
+                network: bitcoin::Network::Bitcoin,
+                node_url: Url::parse("http://example.com").unwrap(),
+            },
+            Bitcoin {
+                network: bitcoin::Network::Testnet,
+                node_url: Url::parse("http://example.com").unwrap(),
+            },
+            Bitcoin {
+                network: bitcoin::Network::Regtest,
+                node_url: Url::parse("http://example.com").unwrap(),
+            },
+        ];
+
+        let actual = file_contents
+            .into_iter()
+            .map(toml::from_str)
+            .collect::<Result<Vec<Bitcoin>, toml::de::Error>>()
+            .unwrap();
+
+        assert_eq!(actual, expected);
+    }
+
+    #[test]
+    fn network_deserializes_correctly() {
+        let file_contents = vec![
+            r#"
+            listen = ["/ip4/0.0.0.0/tcp/9939"]
+            "#,
+            r#"
+            listen = ["/ip4/0.0.0.0/tcp/9939", "/ip4/127.0.0.1/tcp/9939"]
+            "#,
+        ];
+
+        let expected = vec![
+            Network {
+                listen: vec!["/ip4/0.0.0.0/tcp/9939".parse().unwrap()],
+            },
+            Network {
+                listen: (vec![
+                    "/ip4/0.0.0.0/tcp/9939".parse().unwrap(),
+                    "/ip4/127.0.0.1/tcp/9939".parse().unwrap(),
+                ]),
+            },
+        ];
+
+        let actual = file_contents
+            .into_iter()
+            .map(toml::from_str)
+            .collect::<Result<Vec<Network>, toml::de::Error>>()
+            .unwrap();
+
+        assert_eq!(actual, expected);
+    }
+}
